fix bias issue

+++ test/3x/torch/test_transformers.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 import unittest
 
@@ -12,30 +13,47 @@ from neural_compressor.transformers import GPTQConfig, RtnConfig
 class TestQuantizationConfig(unittest.TestCase):
     @classmethod
     def setUpClass(self):
-        self.model_name = "TheBlokeAI/Mixtral-tiny-GPTQ"
+        self.model_name = "hf-internal-testing/tiny-random-gptj"
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name)
         self.prompt = "One day, the little girl"
         self.input_ids = self.tokenizer(self.prompt, return_tensors="pt")["input_ids"]
 
     @classmethod
     def tearDownClass(self):
-        shutil.rmtree("tmp_gptq")
-        shutil.rmtree("tmp_rtn")
+        if os.path.exists("tmp_gptq") and os.path.isdir("tmp_gptq"):
+            shutil.rmtree("tmp_gptq")
+        if os.path.exists("tmp_rtn") and os.path.isdir("tmp_rtn"):
+            shutil.rmtree("tmp_rtn")
 
     def test_gptq(self):
-        quantization_config = GPTQConfig(bits=4, sym=True, damp_percent=0.01, desc_act=True)
+        quantization_config = GPTQConfig(
+            bits=4,
+            sym=True,
+            damp_percent=0.01,
+            desc_act=True,
+            tokenizer=self.tokenizer,
+            n_samples=20,
+            group_size=8,
+            batch_size=5,
+            seq_len=32,
+            block_size=16,
+        )
         user_model = INCModelForCausalLM.from_pretrained(self.model_name, quantization_config=quantization_config)
         output = user_model(self.input_ids)
         user_model.save_pretrained("tmp_gptq")
         loaded_model = INCModelForCausalLM.from_pretrained("tmp_gptq")
         loaded_output = loaded_model(self.input_ids)
-        assert torch.allclose(output, loaded_output, atol=1e-2), "Compare failed!"
+        assert torch.allclose(output.logits, loaded_output.logits, atol=1e-2), "Compare failed!"
 
     def test_rtn(self):
-        quantization_config = RtnConfig(bits=4)
+        quantization_config = RtnConfig(bits=4, group_size=8, sym=False)
         user_model = INCModelForCausalLM.from_pretrained(self.model_name, quantization_config=quantization_config)
         output = user_model(self.input_ids)
         user_model.save_pretrained("tmp_rtn")
         loaded_model = INCModelForCausalLM.from_pretrained("tmp_rtn")
         loaded_output = loaded_model(self.input_ids)
-        assert torch.allclose(output, loaded_output, atol=1e-2), "Compare failed!"
+        assert torch.allclose(output.logits, loaded_output.logits, atol=1e-2), "Compare failed!"
+
+
+if __name__ == "__main__":
+    unittest.main()
